requirements.txt
Flask==2.1.0
google-cloud-error-reporting==1.6.0
gunicorn==20.1.0
